chore: move `JsWitnessMap` into its own file
