+++ src/foreign_calls.rs
@@ -3,7 +3,7 @@ use acvm::FieldElement;
 use js_sys::JsString;
 use wasm_bindgen::{prelude::wasm_bindgen, JsValue};
 
-use crate::js_transforms::js_value_to_field_element;
+use crate::js_witness_map::js_value_to_field_element;
 
 #[wasm_bindgen(typescript_custom_section)]
 const FOREIGN_CALL_HANDLER: &'static str = r#"
